ditch old stuff and start refactoring packet creation loop
