--- packages/web5/lib/src/dids/did_dht/dns/consts.dart
@@ -1,8 +1,0 @@
-// ignore_for_file: constant_identifier_names
-
-const int QUERY_FLAG = 0;
-const int RESPONSE_FLAG = 1 << 15;
-const int FLUSH_MASK = 1 << 15;
-const int NOT_FLUSH_MASK = ~FLUSH_MASK;
-const int QU_MASK = 1 << 15;
-const int NOT_QU_MASK = ~QU_MASK;
