--- packages/web5/lib/src/dids/did_dht/dns/rdata.dart
@@ -1,3 +1,0 @@
-abstract interface class RData {
-  int encodingLength();
-}
