+++ packages/web5/lib/src/dids/did_dht/dns_packet.dart
@@ -1,14 +1,13 @@
+export './dns_packet/name.dart';
 export './dns_packet/answer.dart';
-export './dns_packet/class.dart';
 export './dns_packet/consts.dart';
 export './dns_packet/header.dart';
-export './dns_packet/name.dart';
 export './dns_packet/opcode.dart';
-export './dns_packet/opt_data.dart';
-export './dns_packet/option_code.dart';
 export './dns_packet/packet.dart';
-export './dns_packet/question.dart';
 export './dns_packet/rcode.dart';
 export './dns_packet/rdata.dart';
+export './dns_packet/opt_data.dart';
 export './dns_packet/txt_data.dart';
-export './dns_packet/type.dart';
+export './dns_packet/question.dart';
+export './dns_packet/record_type.dart';
+export './dns_packet/record_class.dart';
