+++ packages/web5/lib/src/dids/did_dht/dns_packet/answer.dart
@@ -1,42 +1,27 @@
 import 'dart:typed_data';
 
-import 'package:web5/src/dids/did_dht/dns_packet/type.dart';
-import 'package:web5/src/dids/did_dht/dns_packet/name.dart';
-import 'package:web5/src/dids/did_dht/dns_packet/rdata.dart';
-import 'package:web5/src/dids/did_dht/dns_packet/class.dart';
+import 'package:web5/src/dids/did_dht/dns_packet/codec.dart';
+import 'package:web5/src/dids/did_dht/dns_packet/rdata_codecs.dart';
 import 'package:web5/src/dids/did_dht/dns_packet/consts.dart';
+import 'package:web5/src/dids/did_dht/dns_packet/name.dart';
 import 'package:web5/src/dids/did_dht/dns_packet/opt_data.dart';
-import 'package:web5/src/dids/did_dht/dns_packet/txt_data.dart';
-
-class BaseAnswer<D> {
-  BaseAnswer({
-    required this.type,
-    required this.name,
-    this.ttl,
-    this.classType,
-    required this.data,
-  });
-
-  DnsType type;
-  String name;
-  int? ttl;
-  String? classType;
-  D data;
-}
+import 'package:web5/src/dids/did_dht/dns_packet/rdata.dart';
+import 'package:web5/src/dids/did_dht/dns_packet/record_class.dart';
+import 'package:web5/src/dids/did_dht/dns_packet/record_type.dart';
 
 /// Represents an answer section in a DNS packet.
-class DnsAnswer {
+class Answer<T extends RData> {
   /// The domain name to which this resource record pertains.
-  late DnsName name;
+  late RecordName name;
 
   /// The type of the resource record. Specifies the meaning of the data in the RData field.
-  late DnsType type;
+  late RecordType type;
 
   /// The class of the data in the RData field.
-  late DnsClass klass;
+  late RecordClass klass;
 
   /// The specific data for this resource record, according to its type.
-  late RData data;
+  late T data;
 
   /// The Time-To-Live of the resource record. This value is the number of seconds
   /// that the resource record may be cached before it should be discarded.
@@ -63,10 +48,7 @@ class DnsAnswer {
   /// Options for OPT records, dynamically determined based on the specific type of option.
   late dynamic options;
 
-  /// The total number of bytes consumed in decoding this DNS answer.
-  int numBytes = 0;
-
-  DnsAnswer({
+  Answer({
     required this.name,
     required this.type,
     required this.klass,
@@ -81,92 +63,118 @@ class DnsAnswer {
     this.options,
   });
 
-  /// Decodes a [DnsAnswer] from a byte buffer [buf] starting at the given [offset].
+  Answer._();
+
+  static final codec = _AnswerCodec();
+
+  /// Decodes a [Answer] from a byte buffer [buf] starting at the given [offset].
   ///
   /// Throws [FormatException] if the buffer data cannot be decoded into a valid DNS answer.
-  DnsAnswer.decode(Uint8List buf, int offset) {
+  factory Answer.decode(Uint8List buf, int offset) =>
+      codec.decode(buf, offset: offset).value as Answer<T>;
+
+  Uint8List encode({Uint8List? buf, int offset = 0}) =>
+      codec.encode(this, input: buf, offset: offset).value;
+
+  int encodingLength() {
+    return name.encodingLength() + 8 + data.encodingLength();
+  }
+}
+
+class _AnswerCodec implements Codec<Answer> {
+  @override
+  EncodeResult encode(
+    Answer answer, {
+    Uint8List? input,
+    int offset = 0,
+  }) {
+    final buf = input ?? Uint8List(answer.encodingLength());
+    final oldOffset = offset;
+
+    final n = RecordName.codec.encode(answer.name, input: buf, offset: offset);
+    offset += n.offset;
+
+    ByteData.view(buf.buffer).setUint16(offset, answer.type.value, Endian.big);
+
+    if (answer.type == RecordType.OPT) {
+      if (answer.name.value != '.') {
+        throw Exception('OPT name must be root.');
+      }
+      ByteData.view(buf.buffer)
+          .setUint16(offset, answer.udpPayloadSize!, Endian.big);
+
+      buf[offset + 4] = answer.extendedRcode!;
+      buf[offset + 5] = answer.ednsVersion!;
+
+      ByteData.view(buf.buffer)
+          .setUint16(offset + 6, answer.flags ?? 0, Endian.big);
+
+      offset += 8;
+      // TODO: need OptDataCodec here
+      offset += answer.options.encode(buf, offset) as int;
+    } else {
+      final klassValue = answer.flush ? FLUSH_MASK : answer.klass.value;
+      ByteData.view(buf.buffer).setUint16(offset + 2, klassValue, Endian.big);
+
+      ByteData.view(buf.buffer).setUint32(offset + 4, answer.ttl, Endian.big);
+
+      offset += 8;
+
+      final result = RDataCodecs.encode(
+        answer.type,
+        answer.data,
+        input: buf,
+        offset: offset,
+      );
+      offset += result.offset;
+    }
+
+    return EncodeResult(buf, offset - oldOffset);
+  }
+
+  @override
+  DecodeResult<Answer> decode(Uint8List buf, {int offset = 0}) {
     final originalOffset = offset;
 
-    name = DnsName.decode(buf, offset);
-    offset += name.numBytes;
+    final nameResult = RecordName.codec.decode(buf, offset: offset);
+    offset += nameResult.offset;
 
     final byteData = ByteData.sublistView(buf);
 
     final rawType = byteData.getUint16(offset, Endian.big);
-    type = DnsType.fromValue(rawType);
+    final type = RecordType.fromValue(rawType);
     offset += 2;
 
-    if (type == DnsType.OPT) {
-      udpPayloadSize = byteData.getUint16(offset + 2, Endian.big);
-      extendedRcode = byteData.getUint8(offset + 4);
-      ednsVersion = byteData.getUint8(offset + 5);
+    final answer = Answer._();
+    answer.name = nameResult.value;
+    answer.type = type;
 
-      flags = byteData.getUint16(offset + 6, Endian.big);
-      flagDo = (flags! >> 15) & 0x1 == 1;
+    if (type == RecordType.OPT) {
+      answer.udpPayloadSize = byteData.getUint16(offset + 2, Endian.big);
+      answer.extendedRcode = byteData.getUint8(offset + 4);
+      answer.ednsVersion = byteData.getUint8(offset + 5);
 
-      options = DnsOptData.decode(buf, offset + 8);
+      answer.flags = byteData.getUint16(offset + 6, Endian.big);
+      answer.flagDo = (answer.flags! >> 15) & 0x1 == 1;
 
-      offset += (8 + options.numBytes).toInt();
+      answer.options = OptionData.decode(buf, offset + 8);
+
+      offset += (8 + answer.options.numBytes).toInt();
     } else {
       final rawDnsClass = byteData.getUint16(offset, Endian.big);
-      klass = DnsClass.fromValue(rawDnsClass & NOT_FLUSH_MASK);
+      answer.klass = RecordClass.fromValue(rawDnsClass & NOT_FLUSH_MASK);
 
-      flush = (rawDnsClass & FLUSH_MASK) != 0;
+      answer.flush = (rawDnsClass & FLUSH_MASK) != 0;
       offset += 2;
 
-      ttl = byteData.getUint32(offset, Endian.big);
+      answer.ttl = byteData.getUint32(offset, Endian.big);
       offset += 4;
 
-      switch (type) {
-        case DnsType.TXT:
-          data = DnsTxtData.decode(buf, offset);
-          offset += data.numBytes;
-          break;
-        default:
-          break;
-      }
+      final result = RDataCodecs.decode(type, buf, offset: offset);
+      answer.data = result.value;
+      offset += result.offset;
     }
 
-    numBytes = offset - originalOffset;
-  }
-
-  Uint8List encode({Uint8List? buf, int offset = 0}) {
-    buf ??= Uint8List(encodingLength());
-    final oldOffset = offset;
-
-    final n = name.encode(buf: buf, offset: offset);
-    offset += n.elementSizeInBytes;
-
-    ByteData.view(buf.buffer).setUint16(offset, type.value, Endian.big);
-    offset += 2;
-
-    if (type == DnsType.OPT) {
-      if (name.value != '.') {
-        throw Exception('OPT name must be root.');
-      }
-      ByteData.view(buf.buffer).setUint16(offset, udpPayloadSize!, Endian.big);
-      buf[offset + 4] = extendedRcode!;
-      buf[offset + 5] = ednsVersion!;
-      ByteData.view(buf.buffer).setUint16(offset + 6, flags ?? 0, Endian.big);
-
-      offset += 8;
-      offset += options.encode(buf, offset) as int;
-    } else {
-      final klassValue = flush ? FLUSH_MASK : klass.value;
-      ByteData.view(buf.buffer).setUint16(offset, klassValue, Endian.big);
-      offset += 2;
-      ByteData.view(buf.buffer).setUint32(offset, ttl, Endian.big);
-      offset += 4;
-
-      // assume that this is a DnsTxtData?
-      offset += (data as DnsTxtData).encodingLength();
-    }
-
-    numBytes = offset - oldOffset;
-    return buf;
-  }
-
-  int encodingLength() {
-    return name.encodingLength() + 8 + type.numBytes;
+    return DecodeResult(answer, offset - originalOffset);
   }
 }
