+++ packages/web5/lib/src/dids/did_dht/dns_packet/codec.dart
