+++ packages/web5/lib/src/dids/did_dht/dns_packet/opt_data.dart
@@ -1,13 +1,12 @@
 import 'dart:typed_data';
-import 'package:web5/src/dids/did_dht/dns_packet/option_code.dart';
 
-class DnsOptData {
+class OptionData {
   final List<Option> options;
   final int numBytes;
 
-  DnsOptData._(this.options, this.numBytes);
+  OptionData._(this.options, this.numBytes);
 
-  factory DnsOptData.decode(Uint8List buf, int offset) {
+  factory OptionData.decode(Uint8List buf, int offset) {
     final int originalOffset = offset;
     final ByteData byteData = ByteData.sublistView(buf);
 
@@ -23,7 +22,7 @@ class DnsOptData {
     }
 
     final int consumedBytes = offset - originalOffset;
-    return DnsOptData._(options, consumedBytes);
+    return OptionData._(options, consumedBytes);
   }
 
   int get encodingLength {
@@ -32,7 +31,7 @@ class DnsOptData {
 }
 
 class Option {
-  late DnsOptionCode code;
+  late OptionCode code;
   late Uint8List data;
   int numBytes = 0;
 
@@ -50,7 +49,7 @@ class Option {
 
     final rawCode = byteData.getUint16(offset);
     offset += 2;
-    code = DnsOptionCode.fromValue(rawCode);
+    code = OptionCode.fromValue(rawCode);
 
     final len = byteData.getUint16(offset);
     offset += 2;
@@ -59,7 +58,7 @@ class Option {
 
     // Handling specific option types
     switch (code) {
-      case DnsOptionCode.CLIENT_SUBNET:
+      case OptionCode.CLIENT_SUBNET:
         family = byteData.getUint16(offset);
         offset += 2;
 
@@ -76,14 +75,14 @@ class Option {
         offset += len - 4;
         break;
 
-      case DnsOptionCode.TCP_KEEPALIVE:
+      case OptionCode.TCP_KEEPALIVE:
         if (len > 0) {
           timeout = byteData.getUint16(offset);
           offset += 2;
         }
         break;
 
-      case DnsOptionCode.KEY_TAG:
+      case OptionCode.KEY_TAG:
         tags = [];
         for (int i = 0; i < len; i += 2) {
           tags!.add(byteData.getUint16(offset));
@@ -96,5 +95,47 @@ class Option {
     }
 
     numBytes = offset - originalOffset;
+  }
+}
+
+// ignore_for_file: constant_identifier_names
+enum OptionCode {
+  LLQ(1, 'LLQ'),
+  UL(2, 'UL'),
+  NSID(3, 'NSID'),
+  DAU(5, 'DAU'),
+  DHU(6, 'DHU'),
+  N3U(7, 'N3U'),
+  CLIENT_SUBNET(8, 'CLIENT_SUBNET'),
+  EXPIRE(9, 'EXPIRE'),
+  COOKIE(10, 'COOKIE'),
+  TCP_KEEPALIVE(11, 'TCP_KEEPALIVE'),
+  PADDING(12, 'PADDING'),
+  CHAIN(13, 'CHAIN'),
+  KEY_TAG(14, 'KEY_TAG'),
+  DEVICEID(26946, 'DEVICEID'),
+  UNKNOWN(-1, 'UNKNOWN'); // Default for unknown or undefined types
+
+  final int value;
+  final String name;
+
+  const OptionCode(this.value, this.name);
+
+  static OptionCode fromValue(int value) {
+    return OptionCode.values.firstWhere(
+      (opt) => opt.value == value,
+      orElse: () => OptionCode.UNKNOWN, // Default or a suitable fallback
+    );
+  }
+
+  static OptionCode fromName(String name) {
+    if (name.startsWith('OPTION_')) {
+      final value = int.tryParse(name.substring(7)) ?? -1;
+      return fromValue(value);
+    }
+    return OptionCode.values.firstWhere(
+      (opt) => opt.name.toUpperCase() == name.toUpperCase(),
+      orElse: () => OptionCode.UNKNOWN, // Default or a suitable fallback
+    );
   }
 }
