+++ packages/web5/lib/src/dids/did_dht/dns_packet/rdata.dart
@@ -1,7 +1,3 @@
-import 'dart:typed_data';
-
 abstract interface class RData {
-  int get numBytes;
-
-  Uint8List encode();
+  int encodingLength();
 }
