+++ packages/web5/lib/src/dids/did_dht/dns_packet/record_class.dart
