+++ packages/web5/lib/src/dids/did_dht/registered_types.dart
