+++ packages/web5/test/dids/did_dht/dns/txt_data_test.dart
@@ -1,6 +1,6 @@
 import 'package:convert/convert.dart';
 import 'package:test/test.dart';
-import 'package:web5/src/dids/did_dht/dns/txt_data.dart';
+import 'package:web5/src/dids/did_dht/dns_packet/txt_data.dart';
 
 import 'vector.dart';
